Add ./bin/dump tool
This tool allows you to see what serialize output is to assist in debugging serialize output
